requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.3
Werkzeug==2.3.6
python-dotenv==1.0.0